Added my own "janky" energy function, and can display results.
Techinically seamcarving can use any energy function, but it typicaly
uses the HOG descriptor which I was unable to understand.  So instread
I'm using my own which simply sums the difference in RGB values and
normalizes based on the num of neighers the pixel has and the min and
max values of all pixels.