Debug symbols were pointing to tup fuse files.
Looking at debug symbols "nm -a seamcarve", I found that the source
symbols pointed to the tup fuse filesystem.  The solution is to use the
"^c^" tup format, which uses chroot to mask the FUSE indrection.

This forces the user of tup to setuid and have the owner be root.

https://github.com/gittup/tup/issues/195